+++ packages/pangea-sdk/tests/integration/__init__.py
@@ -1,9 +1,11 @@
+from .test_ai_guard import TestAIGuard
 from .test_audit import TestAudit
 from .test_audit_tools import TestAuditTools
 from .test_authn import TestAuthN
 from .test_embargo import TestEmbargo
 from .test_file_scan import TestFileScan
 from .test_intel import TestDomainIntel, TestFileIntel, TestIPIntel, TestURLIntel, TestUserIntel
+from .test_prompt_guard import TestPromptGuard
 from .test_redact import TestRedact
 from .test_sanitize import TestSanitize
 from .test_share import TestShare