+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- AI Guard and Prompt Guard services.
+
 ### Changed
 
 - Clarified what `PangeaConfig.environment` affects.